requirements.txt
click==8.0.3
cycler==0.11.0
fonttools==4.29.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
nltk==3.6.7
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
regex==2022.1.18
six==1.16.0
tqdm==4.62.3
wordcloud==1.8.1
